refactor(bench): Switch to criterion
