+++ Cargo.toml
@@ -43,3 +43,24 @@ handlebars = "3.0"
 serde_derive = "1.0"
 serde_json = "1.0"
 regex = "1.0"
+criterion = "0.3"
+
+[[bench]]
+name = "liquid"
+harness = false
+
+[[bench]]
+name = "handlebars"
+harness = false
+
+[[bench]]
+name = "tera_bigcontext"
+harness = false
+
+[[bench]]
+name = "tera_template"
+harness = false
+
+[[bench]]
+name = "tera_tera"
+harness = false
